Modified debugs

+++ src/main/java/aroze/me/spotifythingy/JoinListener.java
@@ -36,14 +36,14 @@ public class JoinListener implements Listener {
                     .asJson();
 
             String name = response.getBody().getArray().getJSONObject(0).getJSONObject("item").getString("name");
-            int progress = response.getBody().getArray().getJSONObject(0).getInt("progress_ms");
-            int duration = response.getBody().getArray().getJSONObject(0).getJSONObject("item").getInt("duration_ms");
+            double progress = response.getBody().getArray().getJSONObject(0).getInt("progress_ms");
+            double duration = response.getBody().getArray().getJSONObject(0).getJSONObject("item").getInt("duration_ms");
 
 
             e.getPlayer().sendMessage("name: " + name);
             e.getPlayer().sendMessage("progress: " + progress);
             e.getPlayer().sendMessage("duration: " + duration);
-            e.getPlayer().sendMessage("progress%: " + progress/duration);
+            e.getPlayer().sendMessage("progress%: " + (Double) (progress/duration));
 
 
             //progress_ms =
